refactor: refactor props and utils into BaseInputField
affects: @fremtind/jkl-text-input-react
