+++ packages/text-input-react/src/TextField.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Label, SupportLabel } from "@fremtind/jkl-typography-react";
-import { BaseInputProps } from "./utils";
-import { BaseInputField } from "./BaseInputField";
+import { BaseInputField, BaseInputProps } from "./BaseInputField";
 
 interface Props extends BaseInputProps {
     label: string;
